Renamed docker-classic to “com.docker.com”.
