+++ cli/cmd/context/rm.go
@@ -32,10 +32,11 @@ import (
 	"errors"
 	"fmt"
 
+	"github.com/spf13/cobra"
+
 	apicontext "github.com/docker/api/context"
 	"github.com/docker/api/context/store"
 	"github.com/docker/api/multierror"
-	"github.com/spf13/cobra"
 )
 
 type removeOpts struct {
