+++ cli/cmd/version.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/spf13/cobra"
 
-	"github.com/docker/api/cli/dockerclassic"
+	"github.com/docker/api/cli/mobycli"
 )
 
 const cliVersion = "0.1.0"
@@ -19,7 +19,7 @@ func VersionCommand() *cobra.Command {
 		Args:  cobra.MaximumNArgs(0),
 		RunE:  runVersion,
 	}
-	// define flags for backward compatibility with docker-classic
+	// define flags for backward compatibility with com.docker.cli
 	flags := cmd.Flags()
 	flags.String("format", "", "Format the output using the given Go template")
 	flags.String("kubeconfig", "", "Kubernetes config file")
@@ -28,11 +28,11 @@ func VersionCommand() *cobra.Command {
 }
 
 func runVersion(cmd *cobra.Command, args []string) error {
-	versionResult, _ := dockerclassic.ExecSilent(cmd.Context())
+	versionResult, _ := mobycli.ExecSilent(cmd.Context())
 	// we don't want to fail on error, there is an error if the engine is not available but it displays client version info
 	// Still, technically the [] byte versionResult could be nil, just let the original command display what it has to display
 	if versionResult == nil {
-		return dockerclassic.ExecCmd(cmd)
+		return mobycli.ExecCmd(cmd)
 	}
 	var s string = string(versionResult)
 	fmt.Print(strings.Replace(s, "\n Version:", "\n Azure integration  "+cliVersion+"\n Version:", 1))
