+++ cli/mobycli/job_windows.go
@@ -1,4 +1,4 @@
-package dockerclassic
+package mobycli
 
 import (
 	"fmt"
