+++ context/store/storedefault.go
@@ -35,7 +35,7 @@ type endpoint struct {
 
 func dockerDefaultContext() (*DockerContext, error) {
 	// ensure we run this using default context, in current context has been damaged / removed in store
-	cmd := exec.Command("docker-classic", "--context", "default", "context", "inspect", "default")
+	cmd := exec.Command("com.docker.cli", "--context", "default", "context", "inspect", "default")
 	var stdout bytes.Buffer
 	cmd.Stdout = &stdout
 	err := cmd.Run()
